:recycle:refactor: seperate redux file redux index 파일 분리 Related to: #243
