+++ src/pages/MainPage/Main.tsx
@@ -37,7 +37,7 @@ const Main = (): JSX.Element => {
 
   const { data, isLoading, isFetching } = useGetPostsQuery(
     { page, postType, feedType },
-    { refetchOnMountOrArgChange: true },
+    { refetchOnFocus: true, refetchOnMountOrArgChange: true },
   );
 
   useEffect(() => {
