+++ src/store/api/reviewApi.ts
@@ -1,0 +1,7 @@
+import { oneLineReviewApi } from '@features/OneLineReviewDialog/api/oneLineReviewApi';
+
+export const reviewApiReducers = {
+  [oneLineReviewApi.reducerPath]: oneLineReviewApi.reducer,
+};
+
+export const reviewApiMiddleware = [oneLineReviewApi.middleware];
