+++ src/store/slices/feedSlice.ts
@@ -1,0 +1,5 @@
+import feedReducer from '@features/FeedPage/slice/feedSlice';
+
+export const feedReducers = {
+  feed: feedReducer,
+};
