+++ src/store/slices/librarySlice.ts
@@ -1,0 +1,5 @@
+import bookShelvesReducer from '@features/BookShelvesPage/slice/bookShelvesSlice';
+
+export const libraryReducers = {
+  bookshelves: bookShelvesReducer,
+};
